requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
